[BUGFIX] Flexform override settings are not applied in application controllers
