+++ Classes/Controller/JobController.php
@@ -6,6 +6,7 @@ namespace PAGEmachine\Ats\Controller;
  */
 
 use PAGEmachine\Ats\Domain\Model\Job;
+use PAGEmachine\Ats\Service\TyposcriptService;
 use TYPO3\CMS\Core\Utility\ArrayUtility;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
 
@@ -30,29 +31,7 @@ class JobController extends ActionController
     {
         parent::initializeAction();
 
-        $this->settings = $this->mergeFlexFormAndTypoScriptSettings($this->settings);
-    }
-
-    /**
-     * Merges global TypoScript and FlexForm settings depending on config (override, override of empty values).
-     *
-     * @return array
-     */
-    public function mergeFlexFormAndTypoScriptSettings($settings = [])
-    {
-        if (!empty($settings['flexForm']) && intval($settings['flexForm']['override']) == 1) {
-            $overrideSettings = $settings['flexForm'];
-
-            ArrayUtility::mergeRecursiveWithOverrule(
-                $settings,
-                $overrideSettings,
-                true,
-                (intval($overrideSettings['overrideEmptyValues']) == 1 ? true : false)
-            );
-        }
-        unset($settings['flexForm']);
-
-        return $settings;
+        $this->settings = TyposcriptService::getInstance()->mergeFlexFormAndTypoScriptSettings($this->settings);
     }
 
     /**
